requirements.txt
pandas==2.1.4 
numpy==1.26.4
alpaca-py==0.28.0
requests==2.32.3
schedule==1.2.2
yfinance
ta